+++ .vscode/launch.json
@@ -1,0 +1,28 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "java",
+            "name": "Current File",
+            "request": "launch",
+            "mainClass": "${file}"
+        },
+        {
+            "type": "java",
+            "name": "RunSystem",
+            "request": "launch",
+            "mainClass": "Currency_Exchange.RunSystem",
+            "projectName": "SAP-Currency-Rate-1_46dfc883"
+        },
+        {
+            "type": "java",
+            "name": "Test",
+            "request": "launch",
+            "mainClass": "Currency_Exchange.Test",
+            "projectName": "SAP-Currency-Rate-1_46dfc883"
+        }
+    ]
+}
