+++ COMP611_Currency_Assignment/src/Currency_Exchange/Node.java
@@ -9,14 +9,18 @@ import java.util.Map;
 
 /**
  *
- * @author snipi
+ * @author snipi we still testing
  */
 public class Node {
     String name;
+    int x;
+    int y;
     Map<Node, Double> links;
     
-    public Node(String name) {
+    public Node(String name, int x, int y) {
         this.name = name;
+        this.x = x;
+        this.y = y;
         this.links = new HashMap<>();
     }
