+++ COMP611_Currency_Assignment/src/Currency_Exchange/SizeSeven.java
@@ -1,0 +1,24 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package Currency_Exchange;
+
+/**
+ *
+ * @author snipi
+ */
+public class SizeSeven extends Shapes{
+    
+    public SizeSeven() {
+        super(7);
+        
+        super.nodes[0] = new Node("A", 220, 30);
+        super.nodes[1] = new Node("B", 370, 120);
+        super.nodes[2] = new Node("C", 390, 280);
+        super.nodes[3] = new Node("D", 290, 400);
+        super.nodes[4] = new Node("E", 140, 400);
+        super.nodes[5] = new Node("F", 50, 280);
+        super.nodes[6] = new Node("G", 70, 120);
+    }
+}
