+++ COMP611_Currency_Assignment/src/Currency_Exchange/SizeThree.java
@@ -1,0 +1,19 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package Currency_Exchange;
+
+/**
+ *
+ * @author snipi
+ */
+public class SizeThree extends Shapes{
+    
+    public SizeThree() {
+        super(3);
+        super.nodes[0] = new Node("A", 50, 60);
+        super.nodes[1] = new Node("B", 370, 60);
+        super.nodes[2] = new Node("C", 210, 320);
+    }
+}
