--- Rate.java
@@ -1,21 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package Currency_Exchange;
-
-/**
- *
- * @author snipi
- */
-public class Rate {
-    String from;
-    String to;
-    double rate;
-    
-    public Rate(String from, String to, double rate) {
-        this.from = from;
-        this.to = to;
-        this.rate = rate;
-    }
-}
