--- Shapes.java
@@ -1,19 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package Currency_Exchange;
-
-/**
- *
- * @author snipi
- */
-public class Shapes {
-    int size;
-    Node[] nodes;
-    
-    public Shapes(int size) {
-        this.size = size;
-        this.nodes = new Node[this.size];
-    }
-}
