--- SizeEight.java
@@ -1,25 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package Currency_Exchange;
-
-/**
- *
- * @author snipi
- */
-public class SizeEight extends Shapes{
-    
-    public SizeEight() {
-        super(8);
-        
-        super.nodes[0] = new Node("A", 140, 30);
-        super.nodes[1] = new Node("B", 290, 30);
-        super.nodes[2] = new Node("C", 410, 130);
-        super.nodes[3] = new Node("D", 410, 290);
-        super.nodes[4] = new Node("E", 290, 400);
-        super.nodes[5] = new Node("F", 130, 400);
-        super.nodes[6] = new Node("G", 30, 290);
-        super.nodes[7] = new Node("H", 30, 130);
-    }
-}
