--- SizeFive.java
@@ -1,22 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package Currency_Exchange;
-
-/**
- *
- * @author snipi
- */
-public class SizeFive extends Shapes{
-    
-    public SizeFive() {
-        super(5);
-        
-        super.nodes[0] = new Node("A", 220, 60);
-        super.nodes[1] = new Node("B", 380, 190);
-        super.nodes[2] = new Node("C", 330, 380);
-        super.nodes[3] = new Node("D", 120, 380);
-        super.nodes[4] = new Node("E", 60, 190);
-    }
-}
