--- SizeFour.java
@@ -1,21 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package Currency_Exchange;
-
-/**
- *
- * @author snipi
- */
-public class SizeFour extends Shapes{
-    
-    public SizeFour() {
-        super(4);
-        
-        super.nodes[0] = new Node("A", 80, 80);
-        super.nodes[1] = new Node("B", 360, 80);
-        super.nodes[2] = new Node("C", 360, 360);
-        super.nodes[3] = new Node("D", 80, 360);
-    }
-}
