--- SizeSix.java
@@ -1,23 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package Currency_Exchange;
-
-/**
- *
- * @author snipi
- */
-public class SizeSix extends Shapes{
-    
-    public SizeSix() {
-        super(6);
-        
-        super.nodes[0] = new Node("A", 130, 50);
-        super.nodes[1] = new Node("B", 310, 50);
-        super.nodes[2] = new Node("C", 420, 220);
-        super.nodes[3] = new Node("D", 310, 390);
-        super.nodes[4] = new Node("E", 130, 390);
-        super.nodes[5] = new Node("F", 30, 220);
-    }
-}
